+++ examples/06-create-customer.php
@@ -9,8 +9,8 @@ try {
   require '../vendor/autoload.php';
 
   // Configurar tu API Key y autenticación
-  $SECRET_API_KEY = "{llave}";
-  $culqi = new Culqi\Culqi(array('api_key' => $SECRET_API_KEY));
+  $SECRET_KEY = "{SECRET KEY}";
+  $culqi = new Culqi\Culqi(array('api_key' => $SECRET_KEY));
 
   // Creando Cargo a una tarjeta
   $customer = $culqi->Customers->create(