+++ lib/Culqi/Cards.php
@@ -16,7 +16,7 @@ class Cards extends Resource {
     *
     * @return all Cards.
     */
-    public function getList($options = NULL) {
+    public function all($options = NULL) {
         return $this->request("GET", self::URL_CARDS, $api_key = $this->culqi->api_key, $options);
     }