+++ lib/Culqi/Client.php
@@ -11,27 +11,24 @@ use Culqi\Error as Errors;
 class Client {
     public function request($method, $url, $api_key, $data = NULL) {
         try {
+            $url_params = is_array($data) ? '?' . http_build_query($data) : '';
             $headers= array("Authorization" => "Bearer ".$api_key, "Content-Type" => "application/json", "Accept" => "application/json");
             $options = array(
                 'timeout' => 120
             );
             if($method == "GET") {
-                $url_params = is_array($data) ? '?' . http_build_query($data) : '';
                 $response = \Requests::get(Culqi::BASE_URL. $url . $url_params, $headers, $options);
             } else if($method == "POST") {
                 $response = \Requests::post(Culqi::BASE_URL . $url, $headers, json_encode($data), $options);
             } else if($method == "PATCH") {
                 $response = \Requests::patch(Culqi::BASE_URL . $url, $headers, json_encode($data), $options);
             } else if($method == "DELETE") {
-                $response = \Requests::delete(Culqi::BASE_URL, $options);
+                $response = \Requests::delete(Culqi::BASE_URL. $url . $url_params, $headers, $options);
             }
         } catch (\Exception $e) {
             throw new Errors\UnableToConnect();
         }
         if ($response->status_code >= 200 && $response->status_code <= 206) {
-            if ($method == "DELETE") {
-                return $response->status_code == 204 || $response->status_code == 200;
-            }
             return json_decode($response->body);
         }
         if ($response->status_code == 400) {