+++ lib/Culqi/Events.php
@@ -16,7 +16,7 @@ class Events extends Resource {
    *
    * @return all Events.
   */
-  public function getList($options = NULL) {
+  public function all($options = NULL) {
       return $this->request("GET", self::URL_EVENTS, $api_key = $this->culqi->api_key, $options);
   }