+++ lib/Culqi/Refunds.php
@@ -16,7 +16,7 @@ class Refunds extends Resource {
      *
      * @return all Refunds.
      */
-    public function getList($options = NULL) {
+    public function all($options = NULL) {
         return $this->request("GET", self::URL_REFUNDS, $api_key = $this->culqi->api_key, $options);
     }