+++ lib/Culqi/Subscriptions.php
@@ -16,7 +16,7 @@ class Subscriptions extends Resource {
      *
      * @return all Subscriptions.
      */
-    public function getList($options = NULL) {
+    public function all($options = NULL) {
         return $this->request("GET", self::URL_SUBSCRIPTIONS, $api_key = $this->culqi->api_key, $options);
     }