+++ lib/Culqi/Transfers.php
@@ -16,7 +16,7 @@ class Transfers extends Resource {
      *
      * @return all Transfers.
      */
-    public function getList($options = NULL) {
+    public function all($options = NULL) {
         return $this->request("GET", self::URL_TRANSFERS, $api_key = $this->culqi->api_key, $options);
     }